fix(api): capabilities route から SERVER_CAPABILITIES を出し、Next のビルドを直す (#1925)
#1925（PR #1943）が `src/app/api/capabilities/route.ts` に `SERVER_CAPABILITIES` と
`CapabilitiesResponse` を export したため、`npm run build` が

    Type error: Route "src/app/api/capabilities/route.ts" does not match the required
    types of a Next.js Route. "SERVER_CAPABILITIES" is not a valid Route export field.

で落ちていた。App Router の `route.ts` は Next が知っている field（`GET` / `dynamic` /
`revalidate` …）しか export できない。

`src/config/server-capabilities.ts` へ移し、route とテストはそこから import する。
API の応答は 1 バイトも変わらない。

## なぜ着地まで気づかなかったか

- `.commandmate/verify.yaml` に **`build` ゲートが無い**（#1882 が「稼働サーバの足元で
  build すると配信中の成果物を壊す」ため意図的に外している）。
- `npx tsc --noEmit` はこの制約を見ない。Next の route 型チェックは build ステップ内にある。
- CI の Build ジョブは正しく落ちていた（PR #1943 で 10 pass / 1 fail）。
  **オーケストレーターが fail を確認せずマージしたのが直接原因**。

再発防止（ゲート追加の可否）は #1945 で検討する。

Refs #1925, #1943
